requirements.txt
torch==2.3.0
torchaudio==2.3.0
faster-whisper==1.0.2
tts>=0.22.0
moviepy==1.0.3
fire==0.6.0

# Google Services
google-auth==2.29.0
google-api-python-client==2.129.0
google-cloud-translate==3.15.3
